Add Cluster Controller protocol primitives generation. Code refactorings

+++ webapps/smsc_cc/test/ru/novosoft/smsc/admin/closed_groups/TestClosedGroupManager.java
@@ -1,11 +1,8 @@
 package ru.novosoft.smsc.admin.closed_groups;
 
 import ru.novosoft.smsc.admin.AdminException;
-import ru.novosoft.smsc.admin.cluster_controller.ClusterController;
-import ru.novosoft.smsc.admin.filesystem.FileSystem;
 import ru.novosoft.smsc.util.Address;
 
-import java.io.File;
 import java.util.ArrayList;
 
 /**
@@ -26,7 +23,7 @@ public class TestClosedGroupManager extends ClosedGroupManager {
     return g;
   }
 
-  public boolean removeGroup(long groupId) throws AdminException {
+  public boolean removeGroup(int groupId) throws AdminException {
     ClosedGroup group2remove = null;
     for (ClosedGroup g : groups) {
       if (g.getId() == groupId) {
